Move getSequenceForFeatures() method back to SequenceService
