+++ test/integration/org/bbop/apollo/CdsServiceIntegrationSpec.groovy
@@ -58,7 +58,7 @@ class CdsServiceIntegrationSpec extends AbstractIntegrationSpec{
         String getCdsSequenceString = getSequenceString.replaceAll("@UNIQUENAME@", transcript.uniqueName)
         getCdsSequenceString = getCdsSequenceString.replaceAll("@SEQUENCE_TYPE@", FeatureStringEnum.TYPE_CDS.value)
         JSONObject commandObject = JSON.parse(getCdsSequenceString) as JSONObject
-        JSONObject getCDSSequenceReturnObject = requestHandlingService.getSequenceForFeatures(commandObject)
+        JSONObject getCDSSequenceReturnObject = sequenceService.getSequenceForFeatures(commandObject)
         
         then: "we should get the anticipated CDS sequence"
         assert getCDSSequenceReturnObject.residues != null
